Fixed pb in internal helper function .fetchChromLengthsFromEnsembl() used by makeTranscriptDbFromBiomart().
Because the db accessible at ftp://ftp.ensembl.org/pub/current_mysql/
is not guaranteed to be in sync with the ENSEMBL GENES db available
thru biomaRt, .fetchChromLengthsFromEnsembl() was not guaranteed to
retrieve the chromosome lengths for the correct version of the
ENSEMBL GENES db. The function now retrieves this information from
ftp://ftp.ensembl.org/pub/release-NN/mysql/ where NN is the release
number (e.g. 58) of the ENSEMBL GENES db that is currently available
thru biomaRt.


git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/branches/RELEASE_2_6/madman/Rpacks/GenomicFeatures@48716 bc3139a8-67e5-0310-9ffc-ced21a209358